requirements.txt
pandas==2.3.3
SQLAlchemy==2.0.45
python-dotenv==1.2.1
mysql-connector-python==9.5.0